pkg: fseval: move to pkg
fseval was fairly ugly in the global github.com/openSUSE/umoci package
(which should be empty), so move it to pkg/fseval. In the future, we
might want to move it to separate project entirely. This also makes the
import graph slightly less hectic.
