+++ cmd/umoci/repack.go
@@ -25,12 +25,12 @@ import (
 	"time"
 
 	"github.com/apex/log"
-	"github.com/openSUSE/umoci"
 	"github.com/openSUSE/umoci/mutate"
 	"github.com/openSUSE/umoci/oci/cas"
 	"github.com/openSUSE/umoci/oci/casext"
 	igen "github.com/openSUSE/umoci/oci/config/generate"
 	"github.com/openSUSE/umoci/oci/layer"
+	"github.com/openSUSE/umoci/pkg/fseval"
 	"github.com/openSUSE/umoci/pkg/mtreefilter"
 	ispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/pkg/errors"
@@ -152,9 +152,9 @@ func repack(ctx *cli.Context) error {
 		"keywords": MtreeKeywords,
 	}).Debugf("umoci: parsed mtree spec")
 
-	fsEval := umoci.DefaultFsEval
+	fsEval := fseval.DefaultFsEval
 	if meta.MapOptions.Rootless {
-		fsEval = umoci.RootlessFsEval
+		fsEval = fseval.RootlessFsEval
 	}
 
 	log.Info("computing filesystem diff ...")
