+++ pkg/fseval/fseval_default.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package umoci
+package fseval
 
 import (
 	"os"
